fixed an issue where datastreams were ignored and no error message by wrapping args.action() in its own try block - changed url from archive to adc - removed unnecessary code and outputs

+++ dev/deploy.sh
@@ -2,5 +2,7 @@
 #dest=act_plotting_main.py
 #[ $# = 1 ] && dest=$1
 #scp act_plotting_main.py armprod-plot:$dest
+ls
+if [ -f act_plotting_main.py ]; then echo "its there"; else echo "nope"; exit; fi
 scp act_plotting_main.py armprod-plot:
